Create Initial list
The first step consists in creating a list, where an integer number "n" is the label (a string, starting at "n = 1"), and two coordinates (x_n,y_n) account for its position in a cartesian plane. No point can repeat, some sort of control when new ccordinates are entered is necessary. If such coordinates were once listed, the new input must be dropped off. Therefore the total amount of cities or points in the plane is "N", provided that for every n<N and m<N the distance between n and m cannot be null. The code should keep the full set of labeled points and coordinates available for further actions.